Testsand typecheck fix
